removed unnecessary annotations for json property returning objects instead of primitives added null checks in getters

+++ src/main/java/org/nhsrc/domain/CheckpointScore.java
@@ -58,7 +58,7 @@ public class CheckpointScore extends AbstractScoreEntity {
     }
 
     @JsonProperty("facilityAssessmentId")
-    public long _getFacilityAssessmentId() {
+    public Integer _getFacilityAssessmentId() {
         return this.facilityAssessment.getId();
     }
